Popup node field (#505)
* start node fields

* select node

* the mentity enu work, I have to text it..

* refractor

* fixed some bugs

* fixed subject issues

+++ electron/src/renderer/project/tab-menu/wikify-form.tsx
@@ -2,23 +2,24 @@ import React from 'react';
 
 import { IReactionDisposer, reaction } from 'mobx';
 import { Button, Col, Form, Row } from 'react-bootstrap';
-import wikiStore, { Layer } from '../../../data/store';
-import { Cell, CellSelection } from '../../../common/general';
+import wikiStore, { Layer } from '../../data/store';
+import { Cell, CellSelection } from '../../common/general';
 import { AnnotationBlock, EntityFields, QNode, QNodeEntry } from '@/renderer/common/dtos';
 
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faTimes } from '@fortawesome/free-solid-svg-icons'
 
-import EntityForm from '../entity-form';
-import { isValidLabel, checkSelectedAnnotationBlocks } from '../../table/table-utils';
+import CustomNodeForm from './custom-node-form';
+import { isValidLabel, checkSelectedAnnotationBlocks } from '../table/table-utils';
 
 interface WikifyFormProperties {
   selectedCell: Cell;
-  onSelectBlock: (applyToBlock: boolean) => void;
+  onSelectBlock?: (applyToBlock: boolean) => void;
   onChange: (key: string, value?: string, instanceOf?: QNode | undefined, searchProperties?: boolean) => Promise<void>;
-  onSubmit: (qnode: QNode, applyToBlock: boolean) => Promise<void>;
-  onRemove: (qnode: QNode, applyToBlock: boolean) => Promise<void>;
-  onCreateQnode: (entityFields: EntityFields, applyToBlock: boolean) => Promise<void>;
+  onSubmit: (qnode: QNode, applyToBlock?: boolean) => Promise<void>;
+  onRemove?: (qnode: QNode, applyToBlock: boolean) => Promise<void>;
+  onCreateQnode: (entityFields: EntityFields, applyToBlock?: boolean) => Promise<void>;
+  field?: string;
 }
 
 interface WikifyFormState {
@@ -45,17 +46,17 @@ class WikifyForm extends React.Component<WikifyFormProperties, WikifyFormState> 
   constructor(props: WikifyFormProperties) {
     super(props);
 
-    const { selectedCell } = this.props;
-    const selected = wikiStore.layers.qnode.find(selectedCell);
+    const { selectedCell, field } = this.props;
+    const selected = field ? undefined : wikiStore.layers.qnode.find(selectedCell);
 
     const entityFields = {
       is_property: false,
-      label: selectedCell.value ? selectedCell.value : "",
+      label: selectedCell.value && !field ? selectedCell.value : "",
       description: "",
       data_type: "string",
     }
 
-    entityFields.is_property = wikiStore.table.selectedBlock?.role === 'property' ? true : false;
+    entityFields.is_property = field ? field === "property" : (wikiStore.table.selectedBlock?.role === 'property' ? true : false);
 
     let customQnode = false;
     if (selected && isValidLabel(selected.label) && isValidLabel(selected.id.substring(1, selected.id.length))) {
@@ -96,7 +97,8 @@ class WikifyForm extends React.Component<WikifyFormProperties, WikifyFormState> 
 
   onChangeQnodes(qnodes: Layer<QNodeEntry>) {
     const { entityFields } = this.state;
-    const selected = qnodes.find(this.props.selectedCell);
+    const { field } = this.props;
+    const selected = field ? undefined : qnodes.find(this.props.selectedCell);
     let customQnode = false;
     if (selected && isValidLabel(selected.label) && isValidLabel(selected.id.substring(1, selected.id.length))) {
       entityFields.is_property = selected.id.startsWith("P");
@@ -131,7 +133,8 @@ class WikifyForm extends React.Component<WikifyFormProperties, WikifyFormState> 
 
   toggleApplyToBlock() {
     const { applyToBlock } = this.state;
-    this.props.onSelectBlock(!applyToBlock);
+    const { onSelectBlock } = this.props;
+    if (onSelectBlock) onSelectBlock(!applyToBlock);
     this.setState({ applyToBlock: !applyToBlock });
   }
 
@@ -139,10 +142,6 @@ class WikifyForm extends React.Component<WikifyFormProperties, WikifyFormState> 
     this.setState({ qnodes });
   }
 
-  // handleOnFocusSearch() {
-  //   const { search, instanceOf, entityFields } = this.state;
-  //   this.props.onChange('search', search, instanceOf, entityFields.is_property);
-  // }
 
   handleOnChangeSearch(event: any) {
     const { instanceOf, entityFields } = this.state;
@@ -194,7 +193,7 @@ class WikifyForm extends React.Component<WikifyFormProperties, WikifyFormState> 
     const { onRemove, selectedCell } = this.props;
     const { selected, applyToBlock } = this.state;
     if (!selected) { return; }
-    onRemove(selected, applyToBlock);
+    if (onRemove) onRemove(selected, applyToBlock);
     this.setState({
       selected: undefined,
       instanceOf: undefined,
@@ -277,23 +276,28 @@ class WikifyForm extends React.Component<WikifyFormProperties, WikifyFormState> 
                 onClick={this.clearSearch.bind(this)} />
             ) : null}
           </Col>
-          {!entityFields.is_property && (
-            <Col sm="12" md="12">
-              <Form.Label className="text-muted">Instance Of</Form.Label>
-              <Form.Control
-                type="text" size="sm"
-                placeholder="node"
-                value={instanceOfSearch}
-                onChange={(event: any) => { this.handleOnChangeInstanceOfSearch(event) }}
-                disabled={customQnode} />
-              {instanceOfSearch && qnodes.length ? (
-                <FontAwesomeIcon
-                  icon={faTimes}
-                  className="clear-button"
-                  onClick={this.clearInstanceOfSearch.bind(this)} />
-              ) : null}
-            </Col>
-          )}
+          {
+            !entityFields.is_property ? (
+              <Col sm="12" md="12">
+                <Form.Label className="text-muted">Instance Of</Form.Label>
+                <Form.Control
+                  type="text" size="sm"
+                  placeholder="node"
+                  value={instanceOfSearch}
+                  onChange={(event: any) => { this.handleOnChangeInstanceOfSearch(event) }}
+                  disabled={customQnode} />
+                {
+                  instanceOfSearch && qnodes.length ? (
+                    <FontAwesomeIcon
+                      icon={faTimes}
+                      className="clear-button"
+                      onClick={this.clearInstanceOfSearch.bind(this)} />
+                  )
+                    : null
+                }
+              </Col>
+            ) : null
+        }
         </Form.Group>
       )
     }
@@ -359,7 +363,7 @@ class WikifyForm extends React.Component<WikifyFormProperties, WikifyFormState> 
           <Col sm="12" md="12">
             <input id="check-block"
               type="checkbox"
-              defaultChecked={applyToBlock}
+              checked={applyToBlock}
               onChange={this.toggleApplyToBlock.bind(this)} />
             <Form.Label htmlFor="check-block" className="text-muted">Apply to block</Form.Label>
           </Col>
@@ -368,12 +372,12 @@ class WikifyForm extends React.Component<WikifyFormProperties, WikifyFormState> 
     }
   }
 
-  renderEntityForm() {
+  renderCustomNodeForm() {
     const { entityFields, customQnode } = this.state;
     if (customQnode) {
       return (
         <Col sm="12" md="12">
-          <EntityForm
+          <CustomNodeForm
             entityFields={entityFields}
             handleOnChange={(event: KeyboardEvent, key: "label" | "description" | "data_type" | "is_property") => this.handleOnChangeEntity(event, key)}
           />
@@ -408,19 +412,6 @@ class WikifyForm extends React.Component<WikifyFormProperties, WikifyFormState> 
       }
     }
     this.setState({ entityFields: updatedEntityFields });
-    // , () => {
-    //   const { entityFields, customQnode } = this.state;
-    //   if (entityFields.label && customQnode) {
-    //     if (this.timeoutIdCreateQnode) {
-    //       window.clearTimeout(this.timeoutIdCreateQnode);
-    //     }
-    //     this.timeoutIdCreateQnode = window.setTimeout(() => {
-    //       if (isValidLabel(entityFields.label)) {
-    //         this.handleOnSubmit()
-    //       }
-    //     }, 300);
-    //   }
-    // });
   }
 
   onChangeCustomQnode() {
@@ -446,12 +437,14 @@ class WikifyForm extends React.Component<WikifyFormProperties, WikifyFormState> 
           <Col sm="12" md="12">
             <Button
               size="sm"
-              type="submit"
+              type="button"
               variant="outline-dark"
-              disabled={!(selected || (customQnode && isValidLabel(entityFields.label)))}>
-              Submit
+              disabled={!(selected || (customQnode && isValidLabel(entityFields.label)))}
+              onClick={(event) => this.handleOnSubmit(event)}
+            >
+              { this.props.field ? "OK" : "Submit" }
             </Button>
-            {this.renderRemoveButton()}
+            { this.props.field ? null : this.renderRemoveButton() }
           </Col>
         </Form.Group>
       )
@@ -476,21 +469,21 @@ class WikifyForm extends React.Component<WikifyFormProperties, WikifyFormState> 
 
   render() {
     const { customQnode, entityFields, disabledIsProperty } = this.state;
+    const { field } = this.props;
     return (
-      <Form className="container wikify-form"
-        onSubmit={(event: any) => this.handleOnSubmit(event)}>
+      <div className="wikify-form">
         <Form.Group as={Row} style={{ marginTop: "1rem" }}>
           <Form.Check type="checkbox" inline label="Custom?" checked={customQnode} onChange={() => this.onChangeCustomQnode()} />
         </Form.Group>
         <Form.Group as={Row} style={{ marginTop: "1rem" }} className="search-properties"
           onChange={(event: KeyboardEvent) => this.handleOnChangeEntity(event, "is_property")}>
-          <Form.Check id="check-property-search" type="checkbox" inline label="Is property?" checked={entityFields.is_property}
-            disabled={disabledIsProperty} />
+          <Form.Check id="check-property-search" type="checkbox" inline label="Is property?" checked={entityFields.is_property || field==="property"}
+            disabled={disabledIsProperty || !!field} />
         </Form.Group>
         <Form.Group as={Row}>
           <Col sm="12" md="12">
             {customQnode ?
-              this.renderEntityForm()
+              this.renderCustomNodeForm()
               :
               this.renderSearchInputs()
             }
@@ -499,9 +492,11 @@ class WikifyForm extends React.Component<WikifyFormProperties, WikifyFormState> 
           </Col>
         </Form.Group>
         {this.renderSelectedNode()}
-        {this.renderApplyOptions()}
+        {
+          this.props.field ? null : this.renderApplyOptions()
+        }
         {this.renderSubmitButton()}
-      </Form>
+      </div>
     )
   }
 }
